Transactions structs moved
