+++ blockchain/transaction.go
@@ -15,16 +15,7 @@ type Transaction struct {
 	Outputs []TxOutput
 }
 
-type TxOutput struct {
-	Value  int
-	PubKey string
-}
 
-type TxInput struct {
-	ID  []byte
-	Out int
-	Sig string
-}
 
 func (tx *Transaction) SetID() {
 	var encoded bytes.Buffer
@@ -57,13 +48,7 @@ func (tx *Transaction) IsCoinbase() bool {
 	return len(tx.Inputs) == 1 && len(tx.Inputs[0].ID) == 0 && tx.Inputs[0].Out == -1
 }
 
-func (in *TxInput) CanUnlock(data string) bool {
-	return in.Sig == data
-}
 
-func (out *TxOutput) CanBeUnlocked(data string) bool {
-	return out.PubKey == data
-}
 
 func NewTransaction(from, to string, amount int, chain *Blockchain) *Transaction {
 	var inputs []TxInput
